JCM 예외처리 및 AGV 와의 DataManager , HTTPManager 상속 구조 분리
Former-commit-id: cb3cbbf311362dc02a69be50766848970c69a4be

+++ Source/test/Public/AGV/AGVDataTypes.h
@@ -32,10 +32,10 @@ public:
 	int32 r_enc;
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadWrite, Category = "AGV")
-	int32 odo_l;
+	float odo_l;
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadWrite, Category = "AGV")
-	int32 odo_r;
+	float odo_r;
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadWrite, Category = "AGV")
 	float yaw;
